Move param parsing to new file and add exceptions if required properties do not exist
