+++ src/Shimmer.Tests/Shimmer.Tests.csproj
@@ -172,6 +172,7 @@
     <Compile Include="Core\ReleaseEntryTests.cs" />
     <Compile Include="Core\ReleasePackageTests.cs" />
     <Compile Include="Core\UtilityTests.cs" />
+    <Compile Include="Release\ParseCommandsTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StaticHttpServer.cs" />
     <Compile Include="TestHelpers\AssertExtensions.cs" />
@@ -186,6 +187,10 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\CreateReleasePackage\CreateReleasePackage.csproj">
+      <Project>{DBC4253F-26BD-46D9-94E0-1413AB59613D}</Project>
+      <Name>CreateReleasePackage</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Shimmer.Client\Shimmer.Client.csproj">
       <Project>{f23cefd4-f1ba-4c37-a834-89a5e38055dd}</Project>
       <Name>Shimmer.Client</Name>
